# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Recorder API","description":"Recorder records arbitrary data together with the recording timestamp","version":""},"host":"localhost:8080","paths":{"/data":{"get":{"tags":["recorder"],"summary":"list recorder","description":"List lists all recorded datapoints.","operationId":"recorder#list","parameters":[{"name":"service","in":"query","description":"Service that created datapoint.","required":true,"type":"string"},{"name":"name","in":"query","description":"Name is the name of the datapoint.","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"type":"array","items":{"$ref":"#/definitions/DatapointResponseBody"}}}},"schemes":["http"]},"post":{"tags":["recorder"],"summary":"record-data recorder","description":"RecordData creates a new datapoint.","operationId":"recorder#record-data","parameters":[{"name":"RecordDataRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/RecordDataRequestBody"}}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}}},"definitions":{"DatapointResponseBody":{"title":"DatapointResponseBody","type":"object","properties":{"labels":{"type":"object","description":"Labels is an arbitrary set of key/value pairs attached to the event.","example":{"algo":"type2"},"additionalProperties":true},"name":{"type":"string","description":"Name is the name of the datapoint.","example":"duration"},"service":{"type":"string","description":"Service that created datapoint.","example":"lambda"},"value":{"type":"number","description":"Datapoint value.","example":0.9275684271217525,"format":"double"}},"description":"Datapoint describes a single recording datapoint.","example":{"labels":{"algo":"type2"},"name":"duration","service":"lambda","value":0.05354787274024356},"required":["service","name","value"]},"RecordDataRequestBody":{"title":"RecordDataRequestBody","type":"object","properties":{"labels":{"type":"object","description":"Labels is an arbitrary set of key/value pairs attached to the event.","example":{"algo":"type2"},"additionalProperties":true},"name":{"type":"string","description":"Name is the name of the datapoint.","example":"duration"},"service":{"type":"string","description":"Service that created datapoint.","example":"lambda"},"value":{"type":"number","description":"Datapoint value.","example":0.6096750339921485,"format":"double"}},"example":{"labels":{"algo":"type2"},"name":"duration","service":"lambda","value":0.3712800718040916},"required":["service","name","value"]}}}